+++ src/oms_sensemaking/object_standards/object_standards_data_retriever.py
@@ -3,7 +3,7 @@ import logging
 from oms_sdk.generated.generated_graphql_client import (
     AttributeQuery,
     AttributesAttributesData,
-    NodeNode,
+    NodesNodesData,
     RelationshipNodeQuery,
     RelationshipQuery,
     RelationshipsRelationshipsData,
@@ -18,14 +18,11 @@ class ObjectStandardsDataRetriever:
     def retrieve_data_for_grading(
         self,
         oms_crud_tool: OmsCrudTool,
-        node: NodeNode,
+        node: NodesNodesData,
         required_attributes: list[str],
         required_relationships: list[str],
-    ) -> dict[str, list[AttributesAttributesData] | list[RelationshipsRelationshipsData] | None]:
-        data_to_grade: dict[str, list[AttributesAttributesData] | list[RelationshipsRelationshipsData] | None] = {
-            "attributes": None,
-            "relationships": None,
-        }
+    ) -> dict[str, list[AttributesAttributesData] | list[RelationshipsRelationshipsData]]:
+        data_to_grade: dict[str, list[AttributesAttributesData] | list[RelationshipsRelationshipsData]] = {}
         if required_attributes:
             try:
                 attributes_to_grade = oms_crud_tool.get_attributes(