+++ CHANGELOG.md
@@ -5,9 +5,13 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 
 ### Added
+- Base In-Port SM class and controller
+- The Object Standards Sensemaker is now able to publish results to ATOMS
 
 ### Changed
 - Metrics are now handled with a base publisher class and the prometheus metrics publisher inherits the metrics publisher class
+- Removed Object Standard's update event in the filter
+
 ### Fixed
 
 ## [0.27.0] - 2026-03-30